refactor: allocate module for CoWShed
