+++ apps/cowswap-frontend/src/modules/cowShed/containers/RecoverFundsFromProxy/index.tsx
@@ -1,16 +1,13 @@
 import { atom, useAtom } from 'jotai'
-import { useCallback, useEffect, useState } from 'react'
+import { ReactNode, useCallback, useEffect, useState } from 'react'
 
-import IMG_ICON_MINUS from '@cowprotocol/assets/images/icon-minus.svg'
-import IMG_ICON_PLUS from '@cowprotocol/assets/images/icon-plus.svg'
 import { useNativeTokenBalance } from '@cowprotocol/balances-and-allowances'
 import { getCurrencyAddress, getEtherscanLink, getIsNativeToken } from '@cowprotocol/common-utils'
 import { Command } from '@cowprotocol/types'
-import { ButtonPrimary, ExternalLink, Loader, TokenAmount } from '@cowprotocol/ui'
+import { ButtonPrimary, ExternalLink, Loader } from '@cowprotocol/ui'
 import { Currency, CurrencyAmount } from '@uniswap/sdk-core'
 
 import ms from 'ms.macro'
-import SVG from 'react-inlinesvg'
 import useSWR from 'swr'
 
 import { useErrorModal } from 'legacy/hooks/useErrorMessageAndModal'
@@ -28,88 +25,19 @@ import { useTokenContract } from 'common/hooks/useContract'
 import { CurrencySelectButton } from 'common/pure/CurrencySelectButton'
 import { NewModal } from 'common/pure/NewModal'
 
-import { Content, FAQItem, FAQWrapper, ProxyInfo, Title, Wrapper } from './styled'
-import { useRecoverFundsFromProxy } from './useRecoverFundsFromProxy'
+import { Content, ProxyInfo, Title, Wrapper } from './styled'
+
+import { useRecoverFundsFromProxy } from '../../hooks/useRecoverFundsFromProxy'
+import { BalanceToRecover } from '../../pure/BalanceToRecover'
+import { CoWShedFAQ } from '../../pure/CoWShedFAQ'
 
 const BALANCE_UPDATE_INTERVAL = ms`5s`
 const BALANCE_SWR_CFG = { refreshInterval: BALANCE_UPDATE_INTERVAL, revalidateOnFocus: true }
 
 const selectedCurrencyAtom = atom<Currency | undefined>(undefined)
 
-// TODO: Break down this large function into smaller functions
-// TODO: Add proper return type annotation
-// eslint-disable-next-line max-lines-per-function, @typescript-eslint/explicit-function-return-type
-function FAQ({ explorerLink }: { explorerLink: string | undefined }) {
-  const [openItems, setOpenItems] = useState<Record<number, boolean>>({})
-
-  const handleToggle = (index: number) => (e: React.MouseEvent) => {
-    e.preventDefault()
-    setOpenItems((prev) => ({ ...prev, [index]: !prev[index] }))
-  }
-
-  const FAQ_DATA = [
-    {
-      question: 'What is CoW Shed?',
-      answer: (
-        <>
-          <ExternalLink href="https://github.com/cowdao-grants/cow-shed">CoW Shed</ExternalLink> is a helper contract
-          that enhances user experience inside CoW Swap for features like{' '}
-          <ExternalLink href="https://docs.cow.fi/cow-protocol/reference/core/intents/hooks">CoW Hooks</ExternalLink>
-          .
-          <br />
-          <br />
-          This contract is deployed only once per account. This account becomes the only owner. CoW Shed will act as an
-          intermediary account who will do the trading on your behalf.
-          <br />
-          <br />
-          Because this contract holds the funds temporarily, it's possible the funds are stuck in some edge cases. This
-          tool will help you recover your funds.
-        </>
-      ),
-    },
-    {
-      question: 'How do I recover my funds from CoW Shed?',
-      answer: (
-        <>
-          <ol>
-            <li>
-              {explorerLink ? (
-                <ExternalLink href={explorerLink}>Check in the block explorer</ExternalLink>
-              ) : (
-                'Check in block explorer'
-              )}{' '}
-              if your own CoW Shed has any token
-            </li>
-            <li>Select the token you want to recover from CoW Shed</li>
-            <li>Recover!</li>
-          </ol>
-        </>
-      ),
-    },
-  ]
-
-  return (
-    <FAQWrapper>
-      {FAQ_DATA.map((faq, index) => (
-        <FAQItem key={index} open={openItems[index]}>
-          <summary onClick={handleToggle(index)}>
-            {faq.question}
-            <i>
-              <SVG src={openItems[index] ? IMG_ICON_MINUS : IMG_ICON_PLUS} />
-            </i>
-          </summary>
-          {openItems[index] && <div>{faq.answer}</div>}
-        </FAQItem>
-      ))}
-    </FAQWrapper>
-  )
-}
-
-// TODO: Break down this large function into smaller functions
-// TODO: Add proper return type annotation
-// TODO: Reduce function complexity by extracting logic
-// eslint-disable-next-line max-lines-per-function, @typescript-eslint/explicit-function-return-type, complexity
-export function RecoverFundsFromProxy({ onDismiss }: { onDismiss: Command }) {
+// eslint-disable-next-line max-lines-per-function,complexity
+export function RecoverFundsFromProxy({ onDismiss }: { onDismiss: Command }): ReactNode {
   const [selectedCurrency, setSelectedCurrency] = useAtom(selectedCurrencyAtom)
   const [tokenBalance, setTokenBalance] = useState<CurrencyAmount<Currency> | null>(null)
 
@@ -210,17 +138,7 @@ export function RecoverFundsFromProxy({ onDismiss }: { onDismiss: Command }) {
 
               {selectedTokenAddress ? (
                 <>
-                  <p>
-                    Balance to be recovered:
-                    <br />
-                    {tokenBalance ? (
-                      <b>
-                        <TokenAmount amount={tokenBalance} defaultValue="0" tokenSymbol={tokenBalance.currency} />
-                      </b>
-                    ) : isBalanceLoading ? (
-                      <Loader />
-                    ) : null}
-                  </p>
+                  <BalanceToRecover tokenBalance={tokenBalance} isBalanceLoading={isBalanceLoading} />
                   <ButtonPrimary onClick={recoverFunds} disabled={!hasBalance || isTxSigningInProgress}>
                     {isTxSigningInProgress ? (
                       <Loader />
@@ -235,7 +153,7 @@ export function RecoverFundsFromProxy({ onDismiss }: { onDismiss: Command }) {
                 <div></div>
               )}
             </Content>
-            <FAQ explorerLink={explorerLink} />
+            <CoWShedFAQ explorerLink={explorerLink} />
           </>
         )}
       </NewModal>
