+++ apps/cowswap-frontend/src/modules/cowShed/hooks/useCowShedHooks.ts
@@ -3,9 +3,7 @@ import { useWalletInfo } from '@cowprotocol/wallet'
 
 import useSWR from 'swr'
 
-// TODO: Add proper return type annotation
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export function useCowShedHooks() {
+export function useCowShedHooks(): CowShedHooks | undefined {
   const { chainId } = useWalletInfo()
 
   return useSWR([chainId, 'CowShedHooks'], ([chainId]) => {
