+++ apps/cowswap-frontend/src/modules/cowShed/index.ts
@@ -1,0 +1,1 @@
+export { RecoverFundsFromProxy } from './containers/RecoverFundsFromProxy'
